added menu to DB init. Still needs a menu for parsing with spidersToDB
